% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmlogit.R
\name{bmlogit}
\alias{bmlogit}
\title{Constrained Multinominal Logistic Regression}
\usage{
bmlogit(Y, X, target_Y, pop_X, count_X, control = list())
}
\arguments{
\item{Y}{A matrix of binary outcomes where each columns represents a level of the outcome}

\item{X}{A model matrix of covariates. An intercept must be included.}

\item{target_Y}{A vector of proportions that the final probabilities should
align to (with tolerance set in the argument \code{tol_pred} in control). This should
have the same number of elements as the columns of \code{Y}.}

\item{pop_X}{A table of the unique combinations of the \code{X} variables.  Together
with \code{count_X}, this represents the traditional population table to post-stratify on.}

\item{count_X}{A vector of population counts for each of   the possible combinations
of \code{X}.  Values must be ordered to be the same as the rows of \code{pop_X}.}
}
\value{
The function returns a list of class \code{cmlogit} object.
}
\description{
Constrained Multinominal Logistic Regression
}
\examples{
library(dplyr)

# Georgia CCES -- limit to voted
vv_cces <- cces_GA \%>\%
  filter(vv_turnout_gvm == "Voted") \%>\%
  filter(voted_govR \%in\% c(0, 1))

## population data
statewide <- colSums(elec18_GA[, -1])
target_Y  <- statewide[c("abrams", "kemp")] /
 sum(statewide[c("abrams", "kemp")])
raceage_X    <- model.matrix(~ race_age - 1, data = catalist18_GA)
raceage_N  <- catalist18_GA$count


## survey data
Y1 <- model.matrix(~ voted_govR - 1, data = vv_cces)
Y <- cbind(1 - Y1, Y1)
X <- model.matrix(~ race_age, data = vv_cces)

## fit
fit <- bmlogit(
  Y = Y,
  X = X,
  target_Y = target_Y,
  pop_X    = raceage_X,
  count_X  = raceage_N,
  control  = list(intercept = FALSE, tol_pred = 0.01))

}
