% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-description.R
\docType{data}
\name{catalist18_GA}
\alias{catalist18_GA}
\title{Estimated Proportion of demographics in t}
\format{
Proportions
\describe{
\item{race_age}{Race - age combination, which matches up with the variable
of the same variable in \code{cc18_GA}}
\item{prop_electorate}{Proportion of the turnout electorate, reported
by Catalist (see Source link)}
}
}
\source{
http://bit.ly/3aiAJN2
}
\usage{
catalist18_GA
}
\description{
Estimated Proportion of demographics in t
}
\keyword{datasets}
