% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmlogit_weights.R
\name{cbsw_survey}
\alias{cbsw_survey}
\title{Covariate Balancing Propensity Score}
\usage{
cbsw_survey(x_survey, x_target, n_pop, add_intercept = TRUE)
}
\arguments{
\item{x_survey}{A matrix of survey respondents}

\item{x_target}{A vector of population mean (or higher moments).
Length of \code{x_target} should match
the number of columns in \code{x_survey}.}

\item{n_pop}{Population size.}
}
\value{
A vector of normalized weights.
}
\description{
Covariate Balancing Propensity Score
}
\keyword{internal}
