% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc_cces.R
\docType{data}
\name{cc18_GA}
\alias{cc18_GA}
\title{CCES 2018 from Georgia}
\format{
A survey dataset (n = 1925) from the 2018 CCES
\describe{
\item{year, case_id, ...}{Standard CCES variables. See https://doi.org/10.7910/DVN/II2DB6}
\item{cd}{Congressional District}
\item{voted_gov_party}{Post-Election Self-Report Vote for Governor}
\item{intent_gov_party}{Pre-Election Self-Report Vote for Governor}
\item{voted_rep_party}{Pre-Election Self-Report Vote for U.S. House}
\item{vv_turnout_gvm}{Validated Vote for General Election}
\item{voted_govR}{A numerical version of \code{voted_gov_party} where people
who did not answer the question or said they did not vote are given
a 0.5, Republican vote is given a 1 and Democratic vote is given a 0}
\item{intent_govR}{A numerical version of \code{intent_gov_party} where people
who did not answer the question or said they did not vote are given
a 0.5, Republican vote is given a 1 and Democratic vote is given a 0}
...
}
}
\usage{
cc18_GA
}
\description{
CCES 2018 from Georgia
}
\keyword{datasets}
