% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmlogit_summary.R
\name{coef.bmlogit}
\alias{coef.bmlogit}
\title{Coef function}
\usage{
\method{coef}{bmlogit}(object, ...)
}
\arguments{
\item{object}{bmlogit object}

\item{...}{Unnamed parameters currently unused but kept to avoid R CMD CHECK warnigns}
}
\description{
Coef function
}
