% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmlogit_weights.R
\name{ebal_survey}
\alias{ebal_survey}
\title{Entroply balancing weights for survey}
\usage{
ebal_survey(x_survey, x_target, add_intercept = TRUE)
}
\arguments{
\item{x_survey}{A matrix of survey respondents}

\item{x_target}{A vector of population mean (or higher moments).
Length of \code{x_target} should match
the number of columns in \code{x_survey}.}
}
\value{
A vector of normalized weights.
}
\description{
Entroply balancing weights for survey
}
\keyword{internal}
