% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc_elec.R
\docType{data}
\name{elec18_GA}
\alias{elec18_GA}
\title{Georgia Election Results}
\format{
Election counts
\describe{
\item{cd}{Congressional District},
\item{abrams}{Votes for the Democrat, Stacey Abrams},
\item{kemp}{Votes for the Republican, Brian Kemp},
\item{total}{Total number of votes, including third party}
}
}
\source{
http://bit.ly/2LNgqh5
}
\usage{
elec18_GA
}
\description{
Georgia Election Results
}
\keyword{datasets}
