% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmlogit.R
\name{fn_ll}
\alias{fn_ll}
\title{Log-likelihood function}
\usage{
fn_ll(x, dat_list)
}
\arguments{
\item{x}{A vector of parameters}

\item{Y}{A matrix of choice (n by J)}
}
\description{
Log-likelihood function
}
\keyword{internal}
