% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmlogit.R
\name{fn_ll_grad}
\alias{fn_ll_grad}
\title{Gradient of the log-likelihood function}
\usage{
fn_ll_grad(x, dat_list)
}
\arguments{
\item{x}{A vector of parameters}
}
\description{
Gradient of the log-likelihood function
}
\keyword{internal}
