% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmlogit.R
\name{input_check}
\alias{input_check}
\title{Input check
A function to set the default values of \code{control}.}
\usage{
input_check(control)
}
\arguments{
\item{control}{A list of control parameters.}
}
\description{
Input check
A function to set the default values of \code{control}.
}
\keyword{internal}
