% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmlogit_summary.R
\name{predict.cmlogit}
\alias{predict.cmlogit}
\title{Obtain the predicted probability}
\usage{
\method{predict}{cmlogit}(obj, newdata = NULL, counts = NULL, weights = NULL)
}
\arguments{
\item{obj}{An output of \code{cmlogit()} funcion.}

\item{newdata}{A matrix of newdata. If not provided, the insample fit is returned.}

\item{counts}{A vector of population weights corresponding to the rows of \code{newdata}.
If provided \code{predict} function returns the predicted outcome at the aggregate level.
If \code{counts} is provided, this argument will be ignored.}
}
\value{
A matrix (or a vector) of predicted probabilities.
}
\description{
Obtain the predicted probability
}
