% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmlogit.R
\name{set_control_default}
\alias{set_control_default}
\title{Set control default
A function to set the default values of \code{control}.}
\usage{
set_control_default(control)
}
\arguments{
\item{control}{A list of control parameters.}
}
\description{
Set control default
A function to set the default values of \code{control}.
}
\keyword{internal}
