% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmlogit.R
\name{bmlogit}
\alias{bmlogit}
\title{Constrained Multinominal Logistic Regression}
\usage{
bmlogit(Y, X, target_Y, pop_X, count_X, control = list())
}
\arguments{
\item{Y}{A matrix of binary outcomes where each columns represents a level of the outcome}

\item{X}{A model matrix of covariates. An intercept must be included.}

\item{target_Y}{A vector of proportions that the final probabilities should
align to (with tolerance set in the argument \code{tol_pred} in control). This should
have the same number of elements as the columns of \code{Y}.}

\item{pop_X}{A table of the unique combinations of the \code{X} variables.  Together
with \code{count_X}, this represents the traditional population table to post-stratify on.}

\item{count_X}{A vector of population counts for each of   the possible combinations
of \code{X}.  Values must be ordered to be the same as the rows of \code{pop_X}.}

\item{control}{A list of control parameters, which can contain the following items:
\itemize{
  \item \code{max_iter} A integer value that specifies the maximum iterations
  for the EM algorithm. Defaul value is 200.
  \item \code{tol} A tolerance parameter for assessing the convergence. Default value is 1e-5.
  \item \code{mu0} A vector of prior means. The dimension of this parameter should match the number of variables (i.e., \code{ncol(X)}).
   The default value is \code{0}.
  \item \code{Z0} A matrix for the prior variance covariance matrix.
     The dimension of this matrix should match with the dimension of \code{X} i.e., the number of variables).
     The default value is \code{diag(rep(5, ncol(X)))}
  \item \code{verbose} A boolean argument. If set \code{TRUE}, the function shows the log-posterior for each iteration. Default is \code{FALSE}.
  \item \code{intercept} A boolean argument. When \code{X} already contains the intercept term (i.e., a column of ones), this option should be \code{FALSE}.
     Default is \code{TRUE}.
  \item \code{variance} A boolean argument. If \code{FALSE}, \code{emlogit()} skips the variance calculation.
     This is useful when calling \code{emlogit()} from other programs. Default is \code{TRUE}.
  \item \code{initialize} A method for initialization.
      The package currently supports \code{initialize = "random"} (random) or \code{initialize = "ls"} (least square). Default is \code{"ls"}.
  \item \code{initial_value} A matrix of coefficients used for the initial value where the first column should be zeros for identification.
             This matrix should be of dimension K by J where K is the total number of covariates (including the intercept) and J is the total number of choices.
             If supplied, \code{initial_value} overwrites the initial value generated by \code{initialize} option.
}}
}
\value{
The function returns a list of class \code{cmlogit} object.
}
\description{
Constrained Multinominal Logistic Regression
}
\examples{
library(dplyr)

## survey data
Y <- model.matrix(~ educ - 1, data = cc18_GA)
X <- model.matrix(~ age + female + race, data = cc18_GA)[, -1]

## population table
pop_X_df <-  count(acs_race_GA, age, female, race, wt = count, name = "count")
pop_X    <- model.matrix(~ age + female + race, data = pop_X_df)[, -1]
count_X  <- pull(pop_X_df, count)

## population target
edu_tgt  <- count(acs_educ_GA, educ, wt = count, name = "count") \%>\%
  pull(count)

## fit
fit <- bmlogit(
  Y = Y,
  X = X,
  pop_X    = pop_X,
  count_X  = count_X,
  target_Y = edu_tgt,
  control  = list(tol_pred = 0.05))


}
